Header bug fix
Trying to fix a bug where the deployed header acts undesirably when a page is refreshed, but not when navigated to. This update makes "scrollPosition.js" obsolete, but the script will remain temporarily until the bug is confirmed to be gone.
Other easy to spot bugs still exist, but to not interfere with functionality and will be fixed later.